style(resource-binding): cell change data change to named tuple
