+++ packages/models/src/resource-binding/index.ts
@@ -1,4 +1,3 @@
 export * from './manager'
 export * from './utils'
-export * from './interface'
 export * from './types'
